# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mang Herbal e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  slug: string;
  nameCkb: string;
  nameAr: string;
  nameEn: string;
  productCount?: number;
}

export interface Product {
  id: number;
  nameCkb: string;
  nameAr: string;
  nameEn: string;
  descCkb: string;
  descAr: string;
  descEn: string;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  categorySlug: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  badge?: string | null;
  inStock: boolean;
  isFeatured: boolean;
  rating: number;
  reviewCount?: number;
  createdAt: string;
}

export interface ProductInput {
  nameCkb: string;
  nameAr: string;
  nameEn: string;
  descCkb: string;
  descAr: string;
  descEn: string;
  price: number;
  /** @nullable */
  oldPrice?: number | null;
  categorySlug: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  badge?: string | null;
  inStock?: boolean;
  isFeatured?: boolean;
}

export interface ProductUpdate {
  nameCkb?: string;
  nameAr?: string;
  nameEn?: string;
  descCkb?: string;
  descAr?: string;
  descEn?: string;
  price?: number;
  /** @nullable */
  oldPrice?: number | null;
  categorySlug?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  badge?: string | null;
  inStock?: boolean;
  isFeatured?: boolean;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  total: number;
}

export interface CartItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CartItemUpdate {
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  productId: number;
  quantity: number;
  price: number;
  nameCkb: string;
  nameAr: string;
  nameEn: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  userId: number;
  status: OrderStatus;
  total: number;
  /** @nullable */
  customerName?: string | null;
  isSeen: boolean;
  phone: string;
  address: string;
  /** @nullable */
  note?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  name: string;
  phone: string;
  address: string;
  /** @nullable */
  note?: string | null;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface UnseenCountResponse {
  count: number;
}

export interface StoreStats {
  totalProducts: number;
  totalCategories: number;
  totalOrders: number;
  happyCustomers: number;
}

export interface ResetPasswordInput {
  phone: string;
  code: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface ResetCodeResponse {
  code: string;
  expiresAt: string;
}

export interface SiteSetting {
  key: string;
  valueCkb: string;
  valueAr: string;
  valueEn: string;
}

export interface UpdateSettingsInput {
  settings: SiteSetting[];
}

export interface WhatsAppTestInput {
  phone: string;
  apiKey: string;
}

export interface WhatsAppTestResponse {
  success: boolean;
  /** @nullable */
  detail: string | null;
}

export interface AdminUser {
  id: number;
  name: string;
  phone: string;
  role: string;
  createdAt: string;
  hasResetCode: boolean;
  /** @nullable */
  resetCode?: string | null;
  /** @nullable */
  resetCodeExpiresAt?: string | null;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  deliveredOrders: number;
  totalRevenue: number;
  totalUsers: number;
  totalProducts: number;
  totalFavorites: number;
}

export type GetProductsParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
minPrice?: number | null;
/**
 * @nullable
 */
maxPrice?: number | null;
/**
 * @nullable
 */
featured?: boolean | null;
};