# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mang Herbal e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Categories ────────────────────────────────────────────────────────────
...
  /admin/products:
    get:
      operationId: adminGetProducts
      tags: [admin]
      summary: Admin list all products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Admin create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Admin delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using an admin-issued code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Site settings ─────────────────────────────────────────────────────────
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Admin list all site settings including private keys
      responses:
        "200":
          description: All settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteSetting"
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Admin bulk upsert site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteSetting"
  /admin/whatsapp-test:
    post:
      operationId: adminTestWhatsApp
      tags: [admin]
      summary: Send a test WhatsApp message via CallMeBot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsAppTestInput"
      responses:
        "200":
          description: Result of the send attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppTestResponse"
...
  /admin/users/{id}/reset-code:
    post:
      operationId: adminCreateResetCode
      tags: [admin]
      summary: Admin generate a password reset code for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reset code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetCodeResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── Auth ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, admin]
        createdAt:
          type: string

    # ── Category ────────────────────────────────────────────────────────────
...
    Product:
      type: object
      required: [id, nameCkb, nameAr, nameEn, descCkb, descAr, descEn, price, categorySlug, inStock, isFeatured, rating, createdAt]
      properties:
        id:
          type: integer
        nameCkb:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        descCkb:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        price:
          type: integer
        oldPrice:
          type: ["integer", "null"]
        categorySlug:
          type: string
        imageUrl:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [nameCkb, nameAr, nameEn, descCkb, descAr, descEn, price, categorySlug]
      properties:
        nameCkb:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        descCkb:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        price:
          type: integer
        oldPrice:
          type: ["integer", "null"]
        categorySlug:
          type: string
        imageUrl:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        nameCkb:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        descCkb:
          type: string
        descAr:
          type: string
        descEn:
          type: string
        price:
          type: integer
        oldPrice:
          type: ["integer", "null"]
        categorySlug:
          type: string
        imageUrl:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean

    # ── Cart ────────────────────────────────────────────────────────────────
...
    OrderItem:
      type: object
      required: [productId, quantity, price, nameCkb, nameAr, nameEn]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: integer
        nameCkb:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
    Order:
      type: object
      required: [id, userId, status, total, phone, address, items, createdAt, isSeen]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        total:
          type: integer
        customerName:
          type: ["string", "null"]
        isSeen:
          type: boolean
        phone:
          type: string
        address:
          type: string
        note:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
...
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
...
    ResetPasswordInput:
      type: object
      required: [phone, code, newPassword]
      properties:
        phone:
          type: string
        code:
          type: string
        newPassword:
          type: string
          minLength: 6
    ResetCodeResponse:
      type: object
      required: [code, expiresAt]
      properties:
        code:
          type: string
        expiresAt:
          type: string

    # ── Site settings ────────────────────────────────────────────────────────
    SiteSetting:
      type: object
      required: [key, valueCkb, valueAr, valueEn]
      properties:
        key:
          type: string
        valueCkb:
          type: string
        valueAr:
          type: string
        valueEn:
          type: string
    UpdateSettingsInput:
      type: object
      required: [settings]
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/SiteSetting"
    WhatsAppTestInput:
      type: object
      required: [phone, apiKey]
      properties:
        phone:
          type: string
        apiKey:
          type: string
    WhatsAppTestResponse:
      type: object
      required: [success, detail]
      properties:
        success:
          type: boolean
        detail:
          type: ["string", "null"]

    # ── Admin users & stats ──────────────────────────────────────────────────
...