# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mang Herbal e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  status: OrderStatus;
  total: number;
  /** @nullable */
  customerName?: string | null;
  isSeen: boolean;
  phone: string;
  address: string;
  /** @nullable */
  note?: string | null;
  items: OrderItem[];
  createdAt: string;
}