# lib/api-zod/src/generated/types/siteSetting.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mang Herbal e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSetting {
  key: string;
  valueCkb: string;
  valueAr: string;
  valueEn: string;
}