# lib/api-zod/src/generated/types/whatsAppTestInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Mang Herbal e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppTestInput {
  phone: string;
  apiKey: string;
}